Move service to sub package
Reuse readAsset from Globals.
